+++ NEWS.md
@@ -2,7 +2,7 @@
 
 New:
 
-* `fmt_scientific()` can now be used in the `fmt` argument for rounding with scientific notation.
+* `fmt_sci()` can now be used in the `fmt` argument for rounding with scientific notation.
 
 # modelsummary 1.4.0