+++ NEWS.md
@@ -9,6 +9,7 @@ Bugs:
 * Group separators respect `add_rows` with `shape="rbind"`. Thanks to @lrose1 for Report #626.
 * Bad column with horizontal models in `shape` and grouped estimates. Thanks to @iago-pssjd for Report #631.
 * `coef_rename=TRUE` bug with grouped estimates. Thanks to @iago-pssjd for Report #631.
+* Upstream issue #881 in `parameters` meant that `vcov` was no longer used for confidence intervals.
 
 # modelsummary 1.4.0